feat: refactor package structure to organize screens under 'screen' package
