+++ app/src/main/java/com/example/ocr/screen/DrawScreen.kt
@@ -1,4 +1,4 @@
-package com.example.ocr.ui
+package com.example.ocr.screen
 
 import android.graphics.Bitmap
 import android.net.Uri
