+++ app/src/main/java/com/example/ocr/screen/MainScreen.kt
@@ -1,5 +1,6 @@
-package com.example.ocr.ui
+package com.example.ocr.screen
 
+import android.Manifest
 import android.content.pm.PackageManager
 import android.net.Uri
 import androidx.activity.compose.rememberLauncherForActivityResult
@@ -53,12 +54,12 @@ fun MainScreen(onCaptured: (Uri) -> Unit = {}) {
   LaunchedEffect(Unit) {
     val hasPermission = ContextCompat.checkSelfPermission(
       context,
-      android.Manifest.permission.CAMERA
+      Manifest.permission.CAMERA
     ) == PackageManager.PERMISSION_GRANTED
     if (hasPermission) {
       cameraPermissionState.value = true
     } else {
-      permissionLauncher.launch(android.Manifest.permission.CAMERA)
+      permissionLauncher.launch(Manifest.permission.CAMERA)
     }
   }
